requirements.txt
asgiref==3.7.2
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==5.0.1
django-ckeditor==6.7.0
django-cors-headers==4.7.0
django-js-asset==2.2.0
djangorestframework==3.14.0
executing==2.0.1
gunicorn==21.2.0
idna==3.6
ipython==8.20.0
jedi==0.19.1
lxml==5.1.0
matplotlib-inline==0.1.6
packaging==25.0
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyMySQL==1.1.1
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2025.2
urllib3==2.1.0
wcwidth==0.2.13
